backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.1
aiofiles>=23.2.1

# Auth & Security
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-jose>=3.3.0
python-multipart>=0.0.9

# Databases & ORMs
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9  # PostgreSQL
PyMySQL>=1.1.0          # MySQL
motor>=3.4.0            # MongoDB async driver

# Testing
pytest>=8.1.1
pytest-asyncio>=0.23.5
httpx>=0.27.0

# LangChain Ecosystem
langchain>=0.1.14
langchain-core>=0.1.41
langchain-openai>=0.1.4
langchain-community>=0.0.19  # Needed for ChatOpenAI, etc.
langchain-text-splitters>=0.0.1  # Confirmed naming on PyPI
